requirement.txt
langchain
langchain_community
streamlit
python-dotenv
pypdf